Made can't find group line easier to understand
Changed       msg.send "Oops, can't find anyone with 'a #{group} member' role!"
to       msg.send "Oops, can't find anyone with a '#{group}' member role!"

Better reflects emphasis on the variable as it is '#{group}' that is the user input, not 'a #{group} member'.
i.e. A user types
hubot standup for engineering

not

hubot standup for a engineering member

Might seem pedantic but it confused me for a bit while troubleshooting why the 'standup for' command isn't working for me.